fix(time): move nanosleep policy out of clocksource, add EINTR + remainder (#67)
Time-syscall policy (timespec validation, fault-aware user copies, signal
interruption, remainder computation) previously lived in the hardware
clocksource module (driver/time) via sleep_ns/sleep_timespec, and the
dispatcher dereferenced raw userspace pointers. This conflated POSIX sleep
policy with the clocksource: sub-2ms sleeps busy-spun via TSC, signal
interruption returned no -EINTR/remainder, rem was pre-zeroed before knowing
whether interruption occurred, and raw pointers were held across the context
switch.

Root-cause fix: move all time-syscall policy into a new sys/syscall/time.rs
that drives the existing #66 deadline queue with proper signal/remainder
semantics.

- sys/syscall/time.rs: sys_nanosleep/sys_clock_nanosleep/sys_clock_gettime
  with fault-aware copy_from_user/copy_to_user. Relative sleeps write rem
  only on interruption; absolute sleeps never write rem; past absolute
  deadlines return immediately. Saturating u128 arithmetic (no f64).
  CLOCK_REALTIME<->monotonic deadline translation. Loops on spurious wakes
  re-checking the clock.
- sys/syscall/utils.rs: add copy_from_user/copy_to_user helpers (canonical +
  USER_MAX range check, volatile read/write).
- sys/proc/mod.rs: queue_signal now wakes Sleeping state -> Runnable with
  WakeReason::Signal + clears wait_token/wait_deadline_ns so the stale timer
  entry cannot wake a later wait. Previously only Waiting|SignalWait|AwaitingIo
  were woken.
- driver/time/mod.rs: remove sleep_ns/sleep_timespec/sys_clock_gettime/
  SHORT_SLEEP_THRESHOLD_NS; add realtime_offset_ns() and
  ensure_realtime_offset_inited() accessors. The clocksource module no longer
  touches userspace pointers or implements sleep semantics.
- driver/usb/xhci/xhci.rs: hardware short delay sleep_ns(10) -> timer::wait(10).
- userspace/apps/clockcheck: new sigint mode verifying -EINTR + non-zero
  remainder.
- build.rs: gate kernel linker script on target_os=none so host test builds
  don't get a PT_TLS-less kernel script.
- Fix pre-existing test compile errors in timer/queue.rs (vec! macro) and
  driver/time/hpet.rs (u128 cast).

Regression checks: full time-regression matrix (make test-time) passes across
all 6 cells (tcg core2duo smp1/smp4, tcg qemu64 smp1, tcg max smp1, kvm host
smp1/smp4) with 0 failures and 0 early/backward violations. sigint mode
confirms rc=-1/EINTR with correct non-zero remainder.

+++ kernel/twilight_kernel/src/driver/time/hpet.rs
@@ -272,7 +272,7 @@ mod tests {
         assert!((68..=72).contains(&one_tick), "one_tick={}", one_tick);
         // 1_000_000 ticks ≈ 69.84 ms.
         let million = cycles_to_ns(1_000_000, mult, shift);
-        let expected = 1_000_000u128 * period_fs / 1_000_000;
+        let expected = 1_000_000u128 * period_fs as u128 / 1_000_000;
         assert!((million as i128 - expected as i128).abs() <= 2);
     }
